Improve type inference for destructuring of shaped array assignments
Previously, when analyzing an assignment like `[$a, $b] = expr`, if `expr` had at least one shaped array in its UnionType, we would mostly ignore whatever generic arrays it had in its UnionType (e.g. `Foo[]`) when inferring the type of `$a` and `$b`. That is, if `expr` had a UnionType like `Foo[]|array{0:null,1:null}`, we would only infer that `$a` and `$b` can be `null`.

The goal of this changeset is to instead infer that `$a` and `$b` are `Foo|null`.